# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VentureForge investment sourcing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export interface RegionCount {
  region: string;
  count: number;
  percentage: number;
}

export interface DashboardSummary {
  companyCount: number;
  newThisWeek: number;
  highSignalCount: number;
  trackedDevelopers: number;
  activeRegions: number;
  weeklyDiscovery: TrendPoint[];
  regionBreakdown: RegionCount[];
}

export type ActivityItemKind = typeof ActivityItemKind[keyof typeof ActivityItemKind];


export const ActivityItemKind = {
  company: 'company',
  alert: 'alert',
  scan: 'scan',
  developer: 'developer',
} as const;

/**
 * @nullable
 */
export type ActivityItemSeverity = typeof ActivityItemSeverity[keyof typeof ActivityItemSeverity] | null;


export const ActivityItemSeverity = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface ActivityItem {
  id: string;
  kind: ActivityItemKind;
  title: string;
  description: string;
  timestamp: string;
  /** @nullable */
  companyId: string | null;
  /** @nullable */
  severity: ActivityItemSeverity;
}

export interface Founder {
  name: string;
  title: string;
  /** @nullable */
  linkedinHandle: string | null;
}

export interface BehavioralMetrics {
  builderResilienceScore: number;
  problemSolvingNotes: string;
  commitVelocitySpike: number;
}

export interface RegionalMetadata {
  geographyRegion: string;
  specificCountry: string;
  businessModel: string;
  sector: string;
  employeeCount: number;
}

export type DiscoveredCompanyLastFundingRound = typeof DiscoveredCompanyLastFundingRound[keyof typeof DiscoveredCompanyLastFundingRound];


export const DiscoveredCompanyLastFundingRound = {
  'Pre-Seed': 'Pre-Seed',
  Seed: 'Seed',
  Series_A: 'Series A',
} as const;

export type DiscoveredCompanySignalLabel = typeof DiscoveredCompanySignalLabel[keyof typeof DiscoveredCompanySignalLabel];


export const DiscoveredCompanySignalLabel = {
  High_signal: 'High signal',
  Promising: 'Promising',
  Watch: 'Watch',
} as const;

export interface DiscoveredCompany {
  id: string;
  companyName: string;
  website: string;
  foundersList: Founder[];
  productDescription: string;
  lastFundingRound: DiscoveredCompanyLastFundingRound;
  fundingAmountEstimated: number;
  fundingCurrency: string;
  knownInvestors: string[];
  behavioralMetrics: BehavioralMetrics;
  regionalMetadata: RegionalMetadata;
  signalScore: number;
  signalLabel: DiscoveredCompanySignalLabel;
  discoveredAt: string;
  source: string;
  logoLetter: string;
}

export interface FilterParams {
  search: string;
  region: string[];
  country: string[];
  stage: string[];
  businessModel: string[];
  sector: string[];
}

export interface SavedSearch {
  id: string;
  name: string;
  filterParams: FilterParams;
  createdAt: string;
  matchCount: number;
  isActive: boolean;
}

export interface SavedSearchInput {
  /** @minLength 1 */
  name: string;
  filterParams: FilterParams;
}

export type PortfolioSourceStatus = typeof PortfolioSourceStatus[keyof typeof PortfolioSourceStatus];


export const PortfolioSourceStatus = {
  Pending: 'Pending',
  Scanned: 'Scanned',
  Error: 'Error',
} as const;

export interface PortfolioSource {
  id: string;
  urlLink: string;
  status: PortfolioSourceStatus;
  companyCount: number;
  createdAt: string;
}

export interface PortfolioSourceInput {
  /** @minLength 1 */
  urlLink: string;
}

export type ProviderName = typeof ProviderName[keyof typeof ProviderName];


export const ProviderName = {
  github: 'github',
  stripe: 'stripe',
  notion: 'notion',
  google_sheets: 'google_sheets',
  slack: 'slack',
} as const;

export type IntegrationConnectionCategory = typeof IntegrationConnectionCategory[keyof typeof IntegrationConnectionCategory];


export const IntegrationConnectionCategory = {
  activity: 'activity',
  billing: 'billing',
  delivery: 'delivery',
} as const;

export type IntegrationConnectionStatus = typeof IntegrationConnectionStatus[keyof typeof IntegrationConnectionStatus];


export const IntegrationConnectionStatus = {
  connected: 'connected',
  vault_configured: 'vault_configured',
  not_connected: 'not_connected',
  needs_reauthorization: 'needs_reauthorization',
} as const;

export interface IntegrationConnection {
  provider: ProviderName;
  displayName: string;
  category: IntegrationConnectionCategory;
  status: IntegrationConnectionStatus;
  detail: string;
  /** @nullable */
  lastVerifiedAt: string | null;
}

export type IntegrationVaultInputValues = {[key: string]: string};

export interface IntegrationVaultInput {
  /** @minLength 1 */
  label: string;
  values: IntegrationVaultInputValues;
}

export interface IntegrationVerification {
  provider: ProviderName;
  verified: boolean;
  message: string;
  /** @nullable */
  lastVerifiedAt: string | null;
}

export interface IntegrationDeliveryInput {
  /** @minLength 1 */
  message: string;
}

export interface IntegrationDeliveryResponse {
  provider: ProviderName;
  delivered: boolean;
  message: string;
}

export type BillingStatusTier = typeof BillingStatusTier[keyof typeof BillingStatusTier];


export const BillingStatusTier = {
  Free: 'Free',
  Pro: 'Pro',
  Enterprise: 'Enterprise',
} as const;

export type BillingStatusSubscriptionStatus = typeof BillingStatusSubscriptionStatus[keyof typeof BillingStatusSubscriptionStatus];


export const BillingStatusSubscriptionStatus = {
  active: 'active',
  past_due: 'past_due',
  canceled: 'canceled',
} as const;

export interface BillingStatus {
  tier: BillingStatusTier;
  subscriptionStatus: BillingStatusSubscriptionStatus;
  stripeConnected: boolean;
  hasCustomer: boolean;
  hasSubscription: boolean;
}

export interface BillingPortalResponse {
  url: string;
}

export type SearchQueryParameter = string;

export type RegionQueryParameter = string[];

export type CountryQueryParameter = string[];

export type StageQueryParameter = string[];

export type BusinessModelQueryParameter = string[];

export type SectorQueryParameter = string[];

export type ListCompaniesParams = {
search?: SearchQueryParameter;
region?: RegionQueryParameter;
country?: CountryQueryParameter;
stage?: StageQueryParameter;
businessModel?: BusinessModelQueryParameter;
sector?: SectorQueryParameter;
};