# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VentureForge investment sourcing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List enriched company records
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/RegionQuery"
        - $ref: "#/components/parameters/CountryQuery"
        - $ref: "#/components/parameters/StageQuery"
        - $ref: "#/components/parameters/BusinessModelQuery"
        - $ref: "#/components/parameters/SectorQuery"
      responses:
        "200":
          description: Filtered company records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoveredCompany"
  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company profile
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Company profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveredCompany"
        "404":
          description: Company not found
  /saved-searches:
    get:
      operationId: listSavedSearches
      tags: [saved-searches]
      summary: List saved searches
      responses:
        "200":
          description: Saved searches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedSearch"
    post:
      operationId: createSavedSearch
      tags: [saved-searches]
      summary: Save a sourcing search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedSearchInput"
      responses:
        "201":
          description: Saved search
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
  /saved-searches/{searchId}:
    delete:
      operationId: deleteSavedSearch
      tags: [saved-searches]
      summary: Delete a saved search
      parameters:
        - $ref: "#/components/parameters/SearchId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Saved search not found
...
  /stripe/webhook:
    post:
      operationId: stripeWebhook
      tags: [billing]
      summary: Receive a signed Stripe webhook
      responses:
        "200":
          description: Webhook accepted
        "400":
          description: Invalid signature or payload
components:
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
    SearchId:
      name: searchId
      in: path
      required: true
      schema:
        type: string
...
    SearchQuery:
      name: search
      in: query
      required: false
      schema:
        type: string
    RegionQuery:
      name: region
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    CountryQuery:
      name: country
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    StageQuery:
      name: stage
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    BusinessModelQuery:
      name: businessModel
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    SectorQuery:
      name: sector
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
  schemas:
...
    Founder:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        linkedinHandle:
          type: ["string", "null"]
      required: [name, title, linkedinHandle]
    BehavioralMetrics:
      type: object
      properties:
        builderResilienceScore:
          type: number
        problemSolvingNotes:
          type: string
        commitVelocitySpike:
          type: number
      required: [builderResilienceScore, problemSolvingNotes, commitVelocitySpike]
    RegionalMetadata:
      type: object
      properties:
        geographyRegion:
          type: string
        specificCountry:
          type: string
        businessModel:
          type: string
        sector:
          type: string
        employeeCount:
          type: number
      required: [geographyRegion, specificCountry, businessModel, sector, employeeCount]
    DiscoveredCompany:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        website:
          type: string
        foundersList:
          type: array
          items:
            $ref: "#/components/schemas/Founder"
        productDescription:
          type: string
        lastFundingRound:
          type: string
          enum: [Pre-Seed, Seed, Series A]
        fundingAmountEstimated:
          type: number
        fundingCurrency:
          type: string
        knownInvestors:
          type: array
          items:
            type: string
        behavioralMetrics:
          $ref: "#/components/schemas/BehavioralMetrics"
        regionalMetadata:
          $ref: "#/components/schemas/RegionalMetadata"
        signalScore:
          type: number
        signalLabel:
          type: string
          enum: [High signal, Promising, Watch]
        discoveredAt:
          type: string
        source:
          type: string
        logoLetter:
          type: string
      required:
        - id
        - companyName
        - website
        - foundersList
        - productDescription
        - lastFundingRound
        - fundingAmountEstimated
        - fundingCurrency
        - knownInvestors
        - behavioralMetrics
        - regionalMetadata
        - signalScore
        - signalLabel
        - discoveredAt
        - source
        - logoLetter
    SavedSearch:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        filterParams:
          $ref: "#/components/schemas/FilterParams"
        createdAt:
          type: string
        matchCount:
          type: number
        isActive:
          type: boolean
      required: [id, name, filterParams, createdAt, matchCount, isActive]
    SavedSearchInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        filterParams:
          $ref: "#/components/schemas/FilterParams"
      required: [name, filterParams]
    FilterParams:
      type: object
      properties:
        search:
          type: string
        region:
          type: array
          items:
            type: string
        country:
          type: array
          items:
            type: string
        stage:
          type: array
          items:
            type: string
        businessModel:
          type: array
          items:
            type: string
        sector:
          type: array
          items:
            type: string
      required: [search, region, country, stage, businessModel, sector]
...