# lib/api-zod/src/generated/types/filterParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VentureForge investment sourcing API
 * OpenAPI spec version: 0.1.0
 */

export interface FilterParams {
  search: string;
  region: string[];
  country: string[];
  stage: string[];
  businessModel: string[];
  sector: string[];
}