# lib/api-zod/src/generated/types/regionalMetadata.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VentureForge investment sourcing API
 * OpenAPI spec version: 0.1.0
 */

export interface RegionalMetadata {
  geographyRegion: string;
  specificCountry: string;
  businessModel: string;
  sector: string;
  employeeCount: number;
}